aimms-models/requirements.txt
pandas
scikit-learn
joblib
torch
transformers
xgboost
numpy
fastapi
uvicorn
pydantic
easyocr
opencv-python
python-multipart
Pillow
sentencepiece
protobuf
prophet
